Refactor callback system
