+++ tests/test_callback_unifier.py
@@ -3,12 +3,12 @@ from dash import Input, Output
 
 from core.callback_registry import CallbackRegistry
 from core.plugins.decorators import unified_callback
-from core.unified_callback_coordinator import UnifiedCallbackCoordinator
+from core.truly_unified_callbacks import TrulyUnifiedCallbacks
 
 
 def test_unified_decorator_with_coordinator():
     app = dash.Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @unified_callback(
         coord,
@@ -26,7 +26,7 @@ def test_unified_decorator_with_coordinator():
 
 def test_unified_decorator_method():
     app = dash.Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.unified_callback(
         Output("o", "children"),
