+++ tests/test_unified_callback_coordinator.py
@@ -2,16 +2,12 @@ import pytest
 from dash import Dash, Input, Output
 
 from core.callback_events import CallbackEvent
-from core.callback_manager import CallbackManager
-from core.callback_migration import UnifiedCallbackCoordinatorWrapper
-from core.unified_callback_coordinator import (
-    UnifiedCallbackCoordinator,
-)
+from core.truly_unified_callbacks import TrulyUnifiedCallbacks
 
 
 def test_duplicate_callback_registration():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -36,7 +32,7 @@ def test_duplicate_callback_registration():
 
 def test_output_conflict_detection():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -61,7 +57,7 @@ def test_output_conflict_detection():
 
 def test_allow_duplicate_output():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -86,7 +82,7 @@ def test_allow_duplicate_output():
 
 def test_allow_duplicate_on_output_obj():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -109,7 +105,7 @@ def test_allow_duplicate_on_output_obj():
 
 def test_callback_registration_to_app():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -126,7 +122,7 @@ def test_callback_registration_to_app():
 
 def test_get_callback_conflicts():
     app = Dash(__name__)
-    coord = UnifiedCallbackCoordinator(app)
+    coord = TrulyUnifiedCallbacks(app)
 
     @coord.register_callback(
         Output("out", "children"),
@@ -153,8 +149,7 @@ def test_get_callback_conflicts():
 
 def test_wrapper_event_registration():
     app = Dash(__name__)
-    manager = CallbackManager()
-    wrapper = UnifiedCallbackCoordinatorWrapper(app, manager)
+    wrapper = TrulyUnifiedCallbacks(app)
 
     results = []
 
@@ -162,7 +157,7 @@ def test_wrapper_event_registration():
         results.append(arg)
 
     wrapper.register_event(CallbackEvent.BEFORE_REQUEST, _event, priority=10)
-    manager.trigger(CallbackEvent.BEFORE_REQUEST, 1)
+    wrapper.trigger_event(CallbackEvent.BEFORE_REQUEST, 1)
     assert results == [1]
 
     @wrapper.register_callback(
